Move generate_atom_mapping_1_1 to utils directory
